edited the 'drop_database' function, edtied the 'create_table' function, added the 'list_tables' and 'drop_tables' functions , edited the add_columns' function
added an option in the 'drop_database' function to make sure the user wants to delete the database, edited the 'create_table' function to remove the files created if there was no primary keys chosen, added the 'list_tables' which shows the number of the files that have a corresponding -meta file and to show their names, added the 'drop_table' function which checks if there are files existing in the database and takes input from the user and applies proper validation then delete the files (file.txt and file-meta.txt) which have the same name as the input, edited the 'add_colmns' function to  have proper validation and ask the user if the column is a primary key then stops once a column gets assigned and automatically assigns the rest of the columns to not primary

+++ main-menu.sh
@@ -122,7 +122,9 @@ drop_database() {
     fi
 
     while true; do
-        read -p "Enter database name or type 'exit' to cancel: " dbname
+
+        list_databases
+        read -p "Enter database name to drop or type 'exit' to cancel: " dbname
 
         if [ "$dbname" = "exit" ]; then
             echo "Exiting without dropping a database."
@@ -144,6 +146,17 @@ drop_database() {
             echo "Database '$dbname' does not exist. Please enter a valid name or type 'exit' to cancel."
             continue
         fi
+
+        read -p "Are you sure you want to drop database '$dbname'? (yes/no): " confirm
+        while [[ "$confirm" != "yes" && "$confirm" != "no" ]]; do
+            read -p "Invalid input. Please type 'yes' to confirm or 'no' to cancel: " confirm
+        done
+
+        if [ "$confirm" != "yes" ]; then
+            echo "Dropping database '$dbname' canceled."
+            return
+        fi
+
 
         rm -rf "Databases/$dbname"
         echo "Database '$dbname' dropped successfully."
